+++ packages/go_router/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 15.3.0
+
+- Adds new top level `onEnter` callback with access to current and next route states.
+- Deprecates top level `redirect` in favor of `onEnter`.
+
 ## 15.2.0
 
 - `GoRouteData` now defines `.location`, `.go(context)`, `.push(context)`, `.pushReplacement(context)`, and `replace(context)` to be used for [Type-safe routing](https://pub.dev/documentation/go_router/latest/topics/Type-safe%20routes-topic.html). **Requires go_router_builder >= 3.0.0**.
@@ -6,7 +11,6 @@
 
 * Updates minimum supported SDK version to Flutter 3.27/Dart 3.6.
 * Fixes typo in API docs.
-
 
 ## 15.1.2
 
@@ -30,7 +34,7 @@
 ## 14.8.1
 
 - Secured canPop method for the lack of matches in routerDelegate's configuration.
- 
+
 ## 14.8.0
 
 - Adds `preload` parameter to `StatefulShellBranchData.$branch`.